==> 01_dataframes.R <==
# Datafrmae creation
rm(list = ls())
library(lubridate)
library(dplyr)

# Data importation
data_dir <- 'C:/Users/jumar/OneDrive/Escritorio/N3/datos confe'

files <- list.files(data_dir, full.names = TRUE)

estaciones <- c('EM71', 'P084', 'P018', 'E085', 'R026',
                'EM08', 'R037', 'R062',
                'EM70', 'P021', 'P087', 'P088',
                'A042', 'P023', 'P024', 'R036',
                'A266', 'A126', 'A058', 'A287', 'A288', 'EM12', 'EM07', 'R040', 'E283', 'R038', 'P019', 'P020')

for (i in estaciones){
  
  file <- files[grepl(i, files) & grepl("PQUIN", files)]
  name <- paste0('df.',i)
  cat(i, ':\t',file, '\n')
  assign(name, read.csv(file, skip = 15, header = TRUE, sep = ';'))

  df <- get(name)

  df <- df %>%
    mutate(
      # date and hour
      fecha_hora = dmy_hms(paste(Fecha, Hora), tz = "UTC"),
      # value column as numeric. --- = NA
      value = as.numeric(ifelse(Valor..mm. == "---", NA, gsub(',', '.', Valor..mm.)))
    )

  # no 29 february
  df <- df %>%
    filter(!(month(fecha_hora) == 2 & day(fecha_hora) == 29))

  #rest of  desired columns
  df <- df %>%
    mutate(
      t = year(fecha_hora),
      mes = month(fecha_hora),
      dia.mes = day(fecha_hora),
      h = hour(fecha_hora),
      min = minute(fecha_hora),
      # days from 1-365
      l = {
        l0  <- yday(fecha_hora)
        adj <- if_else(leap_year(fecha_hora) & month(fecha_hora) > 2, 1L, 0L)
        as.integer(l0 - adj)
      }
    ) %>%
    select(t, l, mes, dia.mes, h, min, value)

  names(df)[names(df) == 'value'] <- paste0(i, '.p')

  assign(name, df)
  
}

# final data frame of minutes
df_minutes <- get(paste0('df.', estaciones[1]))

for(i in 2: length(estaciones)){
  df_minutes <- full_join(df_minutes, 
                      get(paste0('df.', estaciones[i])),
                      by = c('t', 'l', 'mes', 'dia.mes', 'h', 'min'))
}

df_minutes <- df_minutes %>%
  arrange(t, mes, dia.mes, h, min) %>% 
  as.data.frame()

save(estaciones, df_minutes,
     file = 'data.RData')

# hourly data frame (sum those in the same hour)
cols_id <- c('t', 'l' ,'mes' , 'dia.mes', 'h')
cols_val <- paste0(estaciones, '.p')

df_hours <- df_minutes %>%
  group_by(across(all_of(cols_id))) %>%
  summarise(
    across(all_of(cols_val), ~sum(.x, na.rm = FALSE)),
    .groups = 'drop'
  ) %>%
  as.data.frame()

save(estaciones, df_minutes, df_hours,
     file = 'data.RData')

#daily data frame (sum those in the same day)
# na.rm = TRUE (at the moment)
cols_id <- c('t', 'l' ,'mes' , 'dia.mes')

df_days <- df_hours %>%
  group_by(across(all_of(cols_id))) %>%
  summarise(
    across(
      all_of(cols_val), 
      ~ if(mean(is.na(.x)) >= 0.25) NA_real_ else sum(.x, na.rm = TRUE)
      ),
    .groups = 'drop'
  ) %>%
  as.data.frame()

save(estaciones, df_minutes, df_hours, df_days,
     file = 'data.RData')


# dias de 8 am a 7.45 (7) al día siguiente
library(dplyr)

df_days_2<- df_hours %>%
  mutate(
    # Ajustar el día según la hora
    l_meteo = ifelse(h < 8, l - 1, l),
    # Si pasamos del día 0 al anterior año
    t_meteo = ifelse(l_meteo == 0, t - 1, t),
    l_meteo = ifelse(l_meteo == 0, 365, l_meteo)  # suponiendo año no bisiesto
  ) %>%
  relocate(t_meteo, l_meteo, .after = h)

cols_id <- c('t_meteo', 'l_meteo')
cols_val <- paste0(estaciones, '.p')
df_days_2 <- df_days_2 %>%
  group_by(across(all_of(cols_id))) %>%
  summarise(
    across(
      all_of(cols_val), 
      ~ if(mean(is.na(.x)) >= 0.25) NA_real_ else sum(.x, na.rm = TRUE)
    ),
    .groups = 'drop'
  ) %>%
  as.data.frame()



#add t_meteo and l_meteo to all dataframes
df_hours <- df_hours %>%
  mutate(
    # Ajustar el día según la hora
    l_meteo = ifelse(h < 8, l - 1, l),
    # Si pasamos del día 0 al anterior año
    t_meteo = ifelse(l_meteo == 0, t - 1, t),
    l_meteo = ifelse(l_meteo == 0, 365, l_meteo)  # suponiendo año no bisiesto
  ) %>%
  relocate(t_meteo, l_meteo, .before = t)

df_minutes <- df_minutes %>%
  mutate(
    # Ajustar el día según la hora
    l_meteo = ifelse(h < 8, l - 1, l),
    # Si pasamos del día 0 al anterior año
    t_meteo = ifelse(l_meteo == 0, t - 1, t),
    l_meteo = ifelse(l_meteo == 0, 365, l_meteo)  # suponiendo año no bisiesto
  ) %>%
  relocate(t_meteo, l_meteo, .before = t)

save(estaciones, df_minutes, df_hours, df_days, df_days_2,
     file = 'data.RData')
